# packages/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "FulfilmentCommandResultDto": {
        "properties": {
          "orderId": {
            "format": "uuid",
            "type": "string"
          },
          "result": {
            "type": "object"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "orderId",
          "status",
          "result"
        ],
        "type": "object"
      },
      "FulfilmentProjectionDto": {
        "properties": {
          "claims": {
            "items": {
              "type": "object"
            },
            "type": "array"
          },
          "jobs": {
            "items": {
              "type": "object"
            },
            "type": "array"
          },
          "orderId": {
            "format": "uuid",
            "type": "string"
          },
          "orderStatus": {
            "type": "string"
          },
          "phase": {
            "type": "object"
          },
          "priceAdjustments": {
            "items": {
              "type": "object"
            },
            "type": "array"
          },
          "replacementRequests": {
            "items": {
              "type": "object"
            },
            "type": "array"
          },
          "shipments": {
            "items": {
              "type": "object"
            },
            "type": "array"
          },
          "slots": {
            "items": {
              "type": "object"
            },
            "type": "array"
          }
        },
        "required": [
          "orderId",
          "orderStatus",
          "phase",
          "jobs",
          "shipments",
          "slots",
          "replacementRequests",
          "claims",
          "priceAdjustments"
        ],
        "type": "object"
      },
...
      "RejectClaimDto": {
        "properties": {
          "reason": {
            "maxLength": 2000,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "opaque",
        "scheme": "bearer",
        "type": "http"
      },
      "operatorSession": {
        "description": "Environment-specific HttpOnly operator session cookie; development uses a local cookie name.",
        "in": "cookie",
        "name": "__Host-taven_admin",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Canonical HTTP contract for Taven clients",
    "title": "Taven API",
    "version": "0.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/orders/{orderId}/fulfilment": {
      "get": {
        "operationId": "OrdersController_get",
        "parameters": [
          {
            "description": "Required for unsafe operator requests. Obtain the session-bound value from GET /admin/auth/session.",
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "maxLength": 256,
              "minLength": 20,
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FulfilmentProjectionDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": "Order was not found"
          }
        },
        "security": [
          {
            "operatorSession": []
          }
        ],
        "summary": "Read the fulfilment and recovery projection",
        "tags": [
          "operator fulfilment"
        ]
      }
    },
    "/admin/orders/{orderId}/fulfilment/adjustments": {
      "post": {
        "operationId": "OrdersController_adjustment",
        "parameters": [
          {
            "description": "Required for unsafe operator requests. Obtain the session-bound value from GET /admin/auth/session.",
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 20,
              "type": "string"
            }
          },
          {
            "description": "Stable command key; replaying altered input returns 409",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "minLength": 8,
              "pattern": "^\\s*\\S[\\s\\S]{6,253}\\S\\s*$",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePriceAdjustmentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FulfilmentCommandResultDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "operatorSession": []
          }
        ],
        "summary": "Record an immutable manual price adjustment",
        "tags": [
          "operator fulfilment"
        ]
      }
    },
    "/admin/orders/{orderId}/fulfilment/adjustments/{adjustmentId}/refund": {
      "post": {
        "operationId": "OrdersController_refundAdjustment",
        "parameters": [
          {
            "description": "Required for unsafe operator requests. Obtain the session-bound value from GET /admin/auth/session.",
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 20,
              "type": "string"
            }
          },
          {
            "description": "Stable command key; replaying altered input returns 409",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "minLength": 8,
              "pattern": "^\\s*\\S[\\s\\S]{6,253}\\S\\s*$",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "adjustmentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FulfilmentCommandResultDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "operatorSession": []
          }
        ],
        "summary": "Request the refund for a manual price adjustment",
        "tags": [
          "operator fulfilment"
        ]
      }
    },
    "/admin/orders/{orderId}/fulfilment/cancel": {
      "post": {
        "operationId": "OrdersController_cancel",
        "parameters": [
          {
            "description": "Required for unsafe operator requests. Obtain the session-bound value from GET /admin/auth/session.",
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 20,
              "type": "string"
            }
          },
          {
            "description": "Stable command key; replaying altered input returns 409",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "minLength": 8,
              "pattern": "^\\s*\\S[\\s\\S]{6,253}\\S\\s*$",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelOrderDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FulfilmentCommandResultDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "operatorSession": []
          }
        ],
        "summary": "Cancel an order before physical handoff and request refunds",
        "tags": [
          "operator fulfilment"
        ]
      }
    },
    "/admin/orders/{orderId}/fulfilment/claim-window-migration": {
      "post": {
        "operationId": "OrdersController_approveLegacyClaimWindow",
        "parameters": [
          {
            "description": "Required for unsafe operator requests. Obtain the session-bound value from GET /admin/auth/session.",
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 20,
              "type": "string"
            }
          },
          {
            "description": "Stable command key; replaying altered input returns 409",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "minLength": 8,
              "pattern": "^\\s*\\S[\\s\\S]{6,253}\\S\\s*$",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveLegacyClaimWindowDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FulfilmentCommandResultDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "operatorSession": []
          }
        ],
        "summary": "Approve the exact Claim window for a legacy accepted Order",
        "tags": [
          "operator fulfilment"
        ]
      }
    },
    "/admin/orders/{orderId}/fulfilment/claims": {
      "post": {
        "operationId": "OrdersController_createClaim",
        "parameters": [
          {
            "description": "Required for unsafe operator requests. Obtain the session-bound value from GET /admin/auth/session.",
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 20,
              "type": "string"
            }
          },
          {
            "description": "Stable command key; replaying altered input returns 409",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "minLength": 8,
              "pattern": "^\\s*\\S[\\s\\S]{6,253}\\S\\s*$",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClaimDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FulfilmentCommandResultDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "operatorSession": []
          }
        ],
        "summary": "Open a manually verified fulfilment Claim",
        "tags": [
          "operator fulfilment"
        ]
      }
    },
    "/admin/orders/{orderId}/fulfilment/claims/{claimId}/refund": {
      "post": {
        "operationId": "OrdersController_refundClaim",
        "parameters": [
          {
            "description": "Required for unsafe operator requests. Obtain the session-bound value from GET /admin/auth/session.",
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 20,
              "type": "string"
            }
          },
          {
            "description": "Stable command key; replaying altered input returns 409",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "minLength": 8,
              "pattern": "^\\s*\\S[\\s\\S]{6,253}\\S\\s*$",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "claimId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FulfilmentCommandResultDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "operatorSession": []
          }
        ],
        "summary": "Credit and refund every unresolved slot in a Claim",
        "tags": [
          "operator fulfilment"
        ]
      }
    },
    "/admin/orders/{orderId}/fulfilment/claims/{claimId}/rejection": {
      "post": {
        "operationId": "OrdersController_rejectClaim",
        "parameters": [
          {
            "description": "Required for unsafe operator requests. Obtain the session-bound value from GET /admin/auth/session.",
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 20,
              "type": "string"
            }
          },
          {
            "description": "Stable command key; replaying altered input returns 409",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "minLength": 8,
              "pattern": "^\\s*\\S[\\s\\S]{6,253}\\S\\s*$",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "claimId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectClaimDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FulfilmentCommandResultDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "description": "Claim has incident, remedy, or financial recovery history"
          }
        },
        "security": [
          {
            "operatorSession": []
          }
        ],
        "summary": "Reject a clean post-delivery quality Claim",
        "tags": [
          "operator fulfilment"
        ]
      }
    },
...
  },
  "servers": [],
...
}